requirements.txt
Flask==3.0.2
pandas==2.2.1
numpy==1.26.4
plotly==5.19.0
requests==2.31.0
python-dotenv==1.0.1
Flask-Caching==2.1.0
scikit-learn==1.3.0
matplotlib==3.7.2
gunicorn==21.2.0
Flask-WTF==1.1.1
Werkzeug==2.3.7
Jinja2==3.1.2 